+++ src/statisticscsv.h
@@ -41,7 +41,7 @@ public:
                                  float psnrY = -1.0,
                                  float psnrU = -1.0,
                                  float psnrV = -1.0,
-                                 float networkLatencyMs = -1.0,
+                                 int64_t networkLatencyMs = -1,
                                  int64_t creationTimestamp = 0.0) override;
   virtual void decodedAudioFrame(QString cname, int64_t timestamp, uint32_t size, uint32_t decodingTime) override;
   virtual void decodedVideoFrame(QString cname, int64_t timestamp, uint32_t size, uint32_t decodingTime, QSize resolution, int64_t e2eLatency) override;
@@ -81,7 +81,7 @@ private:
     uint32_t bandwidthCost = 0;
 
     // One-way network latency estimate in milliseconds for active connections
-    float networkLatencyMs = -1.0f;
+    int64_t networkLatencyMs = -1;
 
     QSize resolution;
     float psnrY = -1.0f;
